// src/app/models/pokemon-types.enum.ts
export enum PokemonType {
    BUG = 'BUG',
    DARK = 'DARK',
    DRAGON = 'DRAGON',
    ELECTRIC = 'ELECTRIC',
    FAIRY = 'FAIRY',
    FIGHTING = 'FIGHTING',
    FIRE = 'FIRE',
    FLYING = 'FLYING',
    GHOST = 'GHOST',
    GRASS = 'GRASS',
    GROUND = 'GROUND',
    ICE = 'ICE',
    NORMAL = 'NORMAL',
    POISON = 'POISON',
    PSYCHIC = 'PSYCHIC',
    ROCK = 'ROCK',
    STEEL = 'STEEL',
    WATER = 'WATER',
}